Reorganizing repo
separating out the dense packer and filemanager into their own files. This will clean up the rexdb.py file as it was getting cumbersome to navigate.
