+++ tests/test_file_manager.py
@@ -1,6 +1,7 @@
 from pyfakefs import fake_filesystem_unittest
 
-from rexdb import RexDB, FileManager
+from rexdb import RexDB
+from src.file_manager import FileManager
 
 
 class FileManagerTest(fake_filesystem_unittest.TestCase):
